# [MKL-DNN] Extended LRN with reusing via Acquire API