Changes discussed here are reimplementing LRN mkl-dnn ops to use common Acquire API . 

Benefits:
- Easier to modify and maintain code e.g. further modifications for Multi-threading and mkl-dnn 1.0 will be easier
- Performance improvement due to Reusing introduced: Google Net v1 (model with LRN) inference via CAPI with bs=1  is **~9%** faster on AVX512 platform e.g. SKX (8180)