@luotao1 2019-07-23
I test Google Net v1 (model with LRN) inference via CAPI with bs=1 on AVX2 platform e.g. E5-2650 v4.
```
./test_analyzer_image_classification --gtest_filter=Analyzer_resnet50.profile_mkldnn --batch_size=1 --warmup --repeat=100 --paddle_num_threads=4 --infer_model=googlenetfluid/ --profile
```
lrn from 241ms to 208ms, speedup 13%
- before
```
Event              Calls       Total       Min.        Max.        Ave.        Ratio.
thread0::conv2d    5700        1881.6      0.085055    3.01909     0.330106    0.787164
**thread0::lrn       200         241.991     0.716039    2.08498     1.20995     0.101236**
thread0::pool2d    1400        178.921     0.0452      0.266267    0.127801    0.0748511
thread0::concat    900         81.7893     0.04769     0.496502    0.090877    0.0342163
thread0::fc        100         4.40745     0.042422    0.046222    0.0440745   0.00184385
thread0::fetch     100         1.21941     0.010564    0.013756    0.0121941   0.000510137
thread0::feed      100         0.427123    0.003667    0.005537    0.00427123  0.000178686
```
- after
```
Event              Calls       Total       Min.        Max.        Ave.        Ratio.
thread0::conv2d    5700        1872.48     0.086678    3.06727     0.328505    0.799478
**thread0::lrn       200         208.1       0.540581    1.58649     1.0405      0.0888512**
thread0::pool2d    1400        178.047     0.046527    0.557823    0.127176    0.0760194
thread0::concat    900         77.3396     0.047206    0.470094    0.0859329   0.0330212
thread0::fc        100         4.48958     0.042211    0.063319    0.0448958   0.00191689
thread0::fetch     100         1.24835     0.011442    0.016381    0.0124835   0.000532998
thread0::feed      100         0.421286    0.003702    0.006086    0.00421286  0.000179874
```