@luotao1 2019-07-23
> ~9% faster

Is this Op-level speedup or Model-level speedup? @jczaja