@jczaja 2019-07-23
@luotao1 Reported speedup is on model level (SKX 8180 AVX512). I stored only partial log:

before:
sample latency: 26.7129 fps: 37.43

...
thread0::lrn: 2524.55
...

After:
sample latency: 24.4794  fps: 40.8506

....
thread0::lrn: 1194.76
......